requirements.txt
fastapi
uvicorn
streamlit
google-genai
python-dotenv
imapclient
email-validator

